Implement method add in exercise 9

+++ Lab_9/src/com/bartoszgorka/HashFunction.java
@@ -1,15 +1,15 @@
 package com.bartoszgorka;
 
 public class HashFunction {
-    private int a;
-    private int b;
+    private long a;
+    private long b;
 
-    public HashFunction(int a, int b) {
+    public HashFunction(long a, long b) {
         this.a = a;
         this.b = b;
     }
 
-    public int hash(int x, int prime) {
-        return (this.a * x + this.b) % prime;
+    public long hash(int x, long prime) {
+        return (this.a * (long)x + this.b) % prime;
     }
 }
